Initial Commit
Initial commit for the cube solver project.  The goal of this project
is to make a program that will solve a 2x2x2 Rubik's cube (AKA the
Pocket Cube) in the shortest number of moves.  The user should be
able to define their cube through a UI that shows a cube, letting the
user drag and drop colors to build their cube and then hit solve, which
will show the needed moves to solve it.  Additionally, we may add a
complete virtual cube where the user can turn faces and everything,
but this will depend on the amount of time the project takes.