Finished cube solving logic, implemented simple UI to test it.
Completely finished the cube_logic class, adding a Cube_Builder class
which builds cubes from symbols, and using it to achieve a significant
speedup in solving, since now only 6 moves have to be used instead
of 12, and only one orientation has to be worried about.  Now, the
cube is solved by assuming that one piece is already in the correct
position, which it turns out is much better.