Begun more complicated GUI, but it is not functional.
Begun work on a more interesting GUI using pygame.  At this point,
a multicolored cube has been drawn, although it still doesn't look
quite right, it is not connected to cube logic in any way, and cannot
be properly interacted with.  Simple rotation has been provided to
view the cube, but it is not right either.