Got the rotation working.
Got the rotation to be based around fixed axes, not dependent on which
way the cube was facing.  This was achieved by orbiting the camera
rather than rotating the cube.  To orbit the camera, its parent was
set to a dummy node at the origin, which could then be rotated, which
gave the exact effect I was hoping for.