Added the ability to click stickers, added Geometry class.
Added the ability to click the cube and get an index of which sticker
is clicked.  However, it only tells an index of the sticker on the
screen; it still has no map of the index to a real sticker depending
on the orientation.  Added the Geometry class and several methods
to help with geometry tasks.