Got the app working fully.
The cube is now good to go.  Colors can be selected, and clicking the
solve button now attempts to solve the cube, printing the solution
if there is one or the appropriate message if the cube is unsolvable.
Additionally, added some good checks in to fast-check for unsolvable
cubes before departing into a huge search space.